# perf: Optimize BTree search using binary search