@ie-Yoshisaur 2024-08-24
> By the way: would you like to spend some more time with this class?

Of course! I’ll work on fixing that issue and submit a PR soon.